Move HardcodedEntitlements to bootstrap package.
This is in preparation for referencing it from initPhase2.
